[MC-128] Rename Dig plugin to DNS
