--- dig/.CHECKSUM
@@ -1,19 +1,0 @@
-{
-	"spec": "7a7596e110e56dd5a588cb784eecc269",
-	"manifest": "95e7a6a89c6fa6b5972a5197a2032da1",
-	"setup": "b7bb50d4add11bebc0235b4f02008864",
-	"schemas": [
-		{
-			"identifier": "forward/schema.py",
-			"hash": "253aef0b2dc378b9b1191020fb5d6429"
-		},
-		{
-			"identifier": "reverse/schema.py",
-			"hash": "9d858df41ce616baba02b78dd75923f6"
-		},
-		{
-			"identifier": "connection/schema.py",
-			"hash": "da5382221ca2a33a2f854e17b068d502"
-		}
-	]
-}
