+++ dns/icon_dns/connection/connection.py
@@ -1,5 +1,6 @@
 import insightconnect_plugin_runtime
 from .schema import ConnectionSchema
+# Custom imports below
 
 
 class Connection(insightconnect_plugin_runtime.Connection):
@@ -8,7 +9,9 @@ class Connection(insightconnect_plugin_runtime.Connection):
         super(self.__class__, self).__init__(input=ConnectionSchema())
 
     def connect(self, params):
+        self.logger.info("Connect: Connecting...")
         pass
 
     def test(self):
-        return {}
+        # TODO: Implement connection test
+        pass
